refactor: fetch the data of the table in the controller for mandatarissen
Refs: lmb-1914
