+++ app/components/mandatarissen/persoon-table.hbs
@@ -1,5 +1,6 @@
 <AuDataTable
   @content={{@content}}
+  @isLoading={{@isLoading}}
   @noDataMessage="Geen mandatarissen gevonden"
   @sort={{@sort}}
   @page={{@page}}
